+++ include/frr/impl/simd.hpp
@@ -6,7 +6,7 @@
 
 namespace frr
 {
-  auto simd(std::uint8_t *const data,
+  auto simd(std::uint32_t *const data,
             const Vector_f64& TL, const Vector_f64& delta,
             const std::size_t max_iteration) -> void;
