+++ include/frr/impl/threads.hpp
@@ -7,7 +7,7 @@
 
 namespace frr
 {
-  auto threads(std::uint8_t *const data,
+  auto threads(std::uint32_t *const data,
                const Vector_f64 &TL, const Vector_f64 &delta,
                const std::size_t max_iteration) -> void;
 }
